+++ src/components/tonieboxes/boxsetup/cc3200/elements/InstallCC3200Tool.tsx
@@ -3,6 +3,7 @@ import { Typography } from "antd";
 import { useTranslation } from "react-i18next";
 import { Link } from "react-router-dom";
 import CodeSnippet from "../../../../common/elements/CodeSnippet";
+import { ExportOutlined } from "@ant-design/icons";
 
 const { Paragraph } = Typography;
 
@@ -17,21 +18,21 @@ export function installCC3200Tool(): JSX.Element {
                 <li>
                     {t("tonieboxes.boxFlashingCommon.installCC3200Tool.pythonText1")}
                     <Link to="https://www.python.org/downloads/" target="_blank">
-                        {t("tonieboxes.boxFlashingCommon.installCC3200Tool.pythonTextLink")}
+                        {t("tonieboxes.boxFlashingCommon.installCC3200Tool.pythonTextLink")} {<ExportOutlined />}
                     </Link>
                     {t("tonieboxes.boxFlashingCommon.installCC3200Tool.pythonText2")}
                 </li>
                 <li>
                     {t("tonieboxes.boxFlashingCommon.installCC3200Tool.gitText1")}
                     <Link to="https://git-scm.com/book/en/v2/Getting-Started-Installing-Git" target="_blank">
-                        {t("tonieboxes.boxFlashingCommon.installCC3200Tool.gitTextLink")}
+                        {t("tonieboxes.boxFlashingCommon.installCC3200Tool.gitTextLink")} {<ExportOutlined />}
                     </Link>
                     {t("tonieboxes.boxFlashingCommon.installCC3200Tool.gitText2")}
                 </li>
                 <li>
                     {t("tonieboxes.boxFlashingCommon.installCC3200Tool.pipText1")}
                     <Link to="https://pip.pypa.io/en/stable/installation/" target="_blank">
-                        {t("tonieboxes.boxFlashingCommon.installCC3200Tool.pipTextLink")}
+                        {t("tonieboxes.boxFlashingCommon.installCC3200Tool.pipTextLink")} {<ExportOutlined />}
                     </Link>
                     {t("tonieboxes.boxFlashingCommon.installCC3200Tool.pipText2")}
                 </li>
@@ -53,7 +54,7 @@ export function installCC3200Tool(): JSX.Element {
             </Paragraph>
             <Paragraph>{t("tonieboxes.boxFlashingCommon.installCC3200Tool.moreInformation")}</Paragraph>
             <Link to="https://github.com/toniebox-reverse-engineering/cc3200tool" target="_blank">
-                {t("tonieboxes.boxFlashingCommon.installCC3200Tool.link")}
+                {t("tonieboxes.boxFlashingCommon.installCC3200Tool.link")} {<ExportOutlined />}
             </Link>
         </>
     );